Move paper version and series to their own file
